Sidebar: Load domain warnings asynchronously
