+++ client/my-sites/current-site/index.jsx
@@ -11,21 +11,18 @@ import { localize } from 'i18n-calypso';
  * Internal dependencies
  */
 import AllSites from 'my-sites/all-sites';
+import AsyncLoad from 'components/async-load';
 import analytics from 'lib/analytics';
 import Button from 'components/button';
 import Card from 'components/card';
 import Site from 'blocks/site';
 import Gridicon from 'gridicons';
-import UpgradesActions from 'lib/upgrades/actions';
-import DomainsStore from 'lib/domains/store';
-import DomainWarnings from 'my-sites/domains/components/domain-warnings';
 import config from 'config';
 import SiteNotice from './notice';
 import CartStore from 'lib/cart/store';
 import { setLayoutFocus } from 'state/ui/layout-focus/actions';
-import { getSelectedSite, getSelectedSiteId } from 'state/ui/selectors';
+import { getSelectedSite } from 'state/ui/selectors';
 import { getCurrentUser } from 'state/current-user/selectors';
-import { isJetpackSite } from 'state/sites/selectors';
 import { getSelectedOrAllSites } from 'state/selectors';
 import { infoNotice, removeNotice } from 'state/notices/actions';
 import { getNoticeLastTimeShown } from 'state/notices/selectors';
@@ -34,44 +31,20 @@ import { recordTracksEvent } from 'state/analytics/actions';
 
 class CurrentSite extends Component {
 	static propTypes = {
-		isJetpack: React.PropTypes.bool,
 		isPreviewShowing: React.PropTypes.bool,
 		siteCount: React.PropTypes.number.isRequired,
 		setLayoutFocus: React.PropTypes.func.isRequired,
-		selectedSiteId: React.PropTypes.number,
 		selectedSite: React.PropTypes.object,
 		translate: React.PropTypes.func.isRequired,
 		anySiteSelected: React.PropTypes.array
 	};
 
-	state = {
-		domainsStore: DomainsStore
-	};
-
 	componentWillMount() {
-		const { selectedSiteId, isJetpack } = this.props;
-		if ( selectedSiteId && ! isJetpack ) {
-			UpgradesActions.fetchDomains( selectedSiteId );
-		}
-
-		DomainsStore.on( 'change', this.handleStoreChange );
 		CartStore.on( 'change', this.showStaleCartItemsNotice );
 	}
 
 	componentWillUnmount() {
-		DomainsStore.off( 'change', this.handleStoreChange );
 		CartStore.off( 'change', this.showStaleCartItemsNotice );
-	}
-
-	componentDidUpdate( prevProps ) {
-		const { selectedSiteId, isJetpack } = this.props;
-		if ( selectedSiteId && ! isJetpack && selectedSiteId !== prevProps.selectedSiteId ) {
-			UpgradesActions.fetchDomains( selectedSiteId );
-		}
-	}
-
-	handleStoreChange = () => {
-		this.setState( { domainsStore: DomainsStore } );
 	}
 
 	showStaleCartItemsNotice = () => {
@@ -101,11 +74,11 @@ class CurrentSite extends Component {
 				}
 			);
 		}
-	}
+	};
 
 	clickStaleCartItemsNotice = () => {
 		this.props.recordTracksEvent( 'calypso_cart_abandonment_notice_click' );
-	}
+	};
 
 	switchSites = ( event ) => {
 		event.preventDefault();
@@ -113,37 +86,7 @@ class CurrentSite extends Component {
 		this.props.setLayoutFocus( 'sites' );
 
 		analytics.ga.recordEvent( 'Sidebar', 'Clicked Switch Site' );
-	}
-
-	getDomainWarnings() {
-		const { selectedSiteId, selectedSite: site } = this.props;
-
-		if ( ! selectedSiteId ) {
-			return null;
-		}
-
-		const domainStore = this.state.domainsStore.getBySite( selectedSiteId );
-		const domains = domainStore && domainStore.list || [];
-
-		return (
-			<DomainWarnings
-				isCompact
-				selectedSite={ site }
-				domains={ domains }
-				position="site-sidebar"
-				ruleWhiteList={ [
-					'unverifiedDomainsCanManage',
-					'unverifiedDomainsCannotManage',
-					'expiredDomainsCanManage',
-					'expiringDomainsCanManage',
-					'expiredDomainsCannotManage',
-					'expiringDomainsCannotManage',
-					'wrongNSMappedDomains',
-					'pendingGappsTosAcceptanceDomains'
-				] }
-			/>
-		);
-	}
+	};
 
 	previewSite = ( event ) => this.props.onClick && this.props.onClick( event );
 
@@ -185,7 +128,7 @@ class CurrentSite extends Component {
 	}
 
 	render() {
-		const { isJetpack, selectedSite, translate, anySiteSelected } = this.props;
+		const { selectedSite, translate, anySiteSelected } = this.props;
 
 		if ( ! anySiteSelected.length ) {
 			return (
@@ -215,6 +158,7 @@ class CurrentSite extends Component {
 						</Button>
 					</span>
 				}
+
 				{ selectedSite
 					? <div>
 						<Site site={ selectedSite } />
@@ -222,7 +166,9 @@ class CurrentSite extends Component {
 					</div>
 					: <AllSites />
 				}
-				{ ! isJetpack && this.getDomainWarnings() }
+
+				<AsyncLoad require="my-sites/current-site/domain-warnings" noPlaceholder={ null } />
+
 				<SiteNotice site={ selectedSite } allSitesPath={ this.props.allSitesPath } />
 			</Card>
 		);
@@ -231,12 +177,9 @@ class CurrentSite extends Component {
 
 export default connect(
 	( state ) => {
-		const selectedSiteId = getSelectedSiteId( state );
 		const user = getCurrentUser( state );
 
 		return {
-			isJetpack: isJetpackSite( state, selectedSiteId ),
-			selectedSiteId,
 			selectedSite: getSelectedSite( state ),
 			anySiteSelected: getSelectedOrAllSites( state ),
 			siteCount: get( user, 'visible_site_count', 0 ),
